+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.0.6 / 2024-04-16
 
+* [FEATURE] allow disabling cache in uri param for hostfact collector
 * [FEATURE] handle scrape timeout and move to dedicated endpoint for host collector
 * [FEATURE] handle scrape timeout and expired-cache for hostfact collector
 * [FIX] return expired cache on scrape error/timeout only if the param expired-cache is true